pyproject.toml
[project]
name = "medical-archive"
version = "0.1.0"
description = ""
authors = [
    {name = "IssacFenton",email = "kiet.dinhnigredo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymongo (>=4.16.0,<5.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
